improve transition into larger datasets
for area-, line- and radar chart

when appending more data, the new data is not transitioned, and the old
data transitions from the edge of end of the chart to e.g. the middle of
the chart.

this fixes so the old dataset transitions smoothly into the whole new dataset.